+++ app/src/ereferrals/res/values/strings.xml
@@ -178,6 +178,7 @@
     <string name="ws_unknown_exception">Unknown exception.</string>
     <string name="feedback_view_image">Open image with...</string>
     <string name="drive_title">Drive root folder</string>
+    <string name="new_survey_disable">The new survey button temporarily disable</string>
 
     <!-- TERMS FROM POEDITOR -->
     <!-- Fuzzy -->
