--- app/src/laos/java/org/eyeseetea/malariacare/fragments/strategies/DashboardUnsentFragmentStrategy.java
@@ -1,10 +1,0 @@
-package org.eyeseetea.malariacare.fragments.strategies;
-
-import android.view.View;
-
-public class DashboardUnsentFragmentStrategy extends ADashboardUnsentFragmentStrategy {
-    @Override
-    public void onCreateView(View rootView) {
-
-    }
-}
