--- app/src/main/java/org/eyeseetea/malariacare/fragments/strategies/ADashboardUnsentFragmentStrategy.java
@@ -1,7 +1,0 @@
-package org.eyeseetea.malariacare.fragments.strategies;
-
-import android.view.View;
-
-public abstract class ADashboardUnsentFragmentStrategy {
-    public abstract void onCreateView(View rootView);
-}
